tcp_westwood : Bump minimum RTT and lower initial RTT duration
Slightly bumping the minimum round-trip-time adds more tolerance
for connections with fluctutating ping to stay connected before
the sender-side TCP stack gives up on the packet and considers
it lost and the connection broken.

Lowering the initial RTT duration works conversely by reducing
tolerance for bad networks so that we avoid getting on bad
connections in the first place.

Change-Id: Ie85b65834693ed40e0a14f0ef0f9b2ca62088354
Signed-off-by: Kneba <abenkenary3@gmail.com>